Removed imlpicit type conversions to not trigger warnings

+++ poly-physx-demo/include/trail_manager.hpp
@@ -22,7 +22,7 @@ namespace ppx_demo
         void write(ini::output &out) const override;
         void read(ini::input &in) override;
 
-        std::size_t p_steps = 150, p_length = 5;
+        std::uint32_t p_steps = 150, p_length = 5;
         float p_line_thickness = 6.f;
         bool p_enabled = false, p_auto_include = true;
